backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.9.1
asyncpg==0.29.0
pysolar==0.11
shapely==2.0.6
httpx==0.27.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
